+++ mod-source-record-manager-server/src/main/java/org/folio/config/ApplicationConfig.java
@@ -1,5 +1,9 @@
 package org.folio.config;
 
+import io.vertx.core.Vertx;
+import org.folio.services.journal.JournalService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
@@ -9,4 +13,12 @@ import org.springframework.context.annotation.Configuration;
   "org.folio.dao",
   "org.folio.services"})
 public class ApplicationConfig {
+
+  @Autowired
+  private Vertx vertx;
+
+  @Bean(value = "journalServiceProxy")
+  public JournalService journalServiceProxy() {
+    return JournalService.createProxy(vertx);
+  }
 }